Added Group preference setting
Preference setting for the user to select the group from which he wants
the translation messages.

P.S. : Currently only the first few groups are added to the list as
creating a list of more than 1000 entries is taking an eternity. Would
be perfer if some limit/offset or pagination was implemented on the
server-side.